refactor: remove the old `Select` component
