+++ src/components/Layer2/Layer2Onboard.tsx
@@ -28,8 +28,7 @@ import { ButtonLink } from "../Buttons"
 import InlineLink from "../Link"
 import OldHeading from "../OldHeading"
 import Text from "../OldText"
-import ReactSelect, { ReactSelectOnChange } from "../ReactSelect"
-import { StyledSelect as Select } from "../SharedStyledComponents"
+import Select, { SelectOnChange } from "../Select"
 
 const Flex50 = (props: ChildOnlyProp) => (
   <Box flex={{ base: "100%", md: "50%" }} {...props} />
@@ -174,7 +173,7 @@ const Layer2Onboard = ({
     },
   } as const
 
-  const handleLayer2SelectChange: ReactSelectOnChange<Layer2Option> = (
+  const handleLayer2SelectChange: SelectOnChange<Layer2Option> = (
     selectedOption
   ) => {
     if (!selectedOption) return
@@ -188,7 +187,7 @@ const Layer2Onboard = ({
     setSelectedL2(selectedOption.l2)
   }
 
-  const handleExchangeOnboard: ReactSelectOnChange<
+  const handleExchangeOnboard: SelectOnChange<
     ExchangeOption | CexOnboardOption
   > = (selectedOption) => {
     if (!selectedOption) return
@@ -242,7 +241,7 @@ const Layer2Onboard = ({
           </Box>
           {/* LeftSelected */}
           <Box mt="auto">
-            <ReactSelect
+            <Select
               instanceId="layer2-left-selected"
               placeholder={t("layer-2-onboard-wallet-input-placeholder")}
               options={layer2Options}
@@ -265,7 +264,7 @@ const Layer2Onboard = ({
           </Box>
           {/* RightSelect */}
           <Box mt="auto">
-            <ReactSelect
+            <Select
               instanceId="exchange-onboard-select"
               options={[
                 {
