+++ src/hooks/useCentralizedExchanges.ts
@@ -12,7 +12,7 @@ import { useTranslation } from "next-i18next"
 // import squarelink from "@/public/wallets/squarelink.png"
 // import trust from "@/public/wallets/trust.png"
 import type { ImageProps } from "@/components/Image"
-import { ReactSelectOnChange } from "@/components/ReactSelect"
+import { SelectOnChange } from "@/components/Select"
 
 import { trackCustomEvent } from "@/lib/utils/matomo"
 
@@ -295,7 +295,7 @@ export const useCentralizedExchanges = () => {
     }))
     .sort((a, b) => a.value.localeCompare(b.value))
 
-  const handleSelectChange: ReactSelectOnChange<ExchangeByCountryOption> = (
+  const handleSelectChange: SelectOnChange<ExchangeByCountryOption> = (
     selectedOption
   ) => {
     if (!selectedOption) return
